+++ PKHeX.Core/PersonalInfo/Table/PersonalTable7.cs
@@ -30,12 +30,12 @@ public sealed class PersonalTable7 : IPersonalTable, IPersonalTable<PersonalInfo
 
     public int GetFormIndex(ushort species, byte form)
     {
-        if ((uint)species <= MaxSpeciesID)
+        if (species <= MaxSpeciesID)
             return Table[species].FormIndex(species, form);
         return 0;
     }
 
-    public bool IsSpeciesInGame(ushort species) => (uint)species <= MaxSpeciesID;
+    public bool IsSpeciesInGame(ushort species) => species <= MaxSpeciesID;
     public bool IsPresentInGame(ushort species, byte form)
     {
         if (!IsSpeciesInGame(species))
